Apply encryption logic
