+++ src/app/components/payment-form.tsx
@@ -1,14 +1,10 @@
-// src/app/components/ManualPaymentForm.tsx
 "use client";
 
-import {
-  encryptCardFieldsIndividually,
-  encryptClientSide,
-} from "@/lib/crypto.client"; // Import our new function
+import { encryptCardFieldsIndividually } from "@/lib/encryption"; // Import our new function
 import type { CardDetails } from "@/types/payment";
 import { FormEvent, useState } from "react";
 
-export default function ManualPaymentForm() {
+export default function PaymentForm() {
   const [status, setStatus] = useState<string>("Ready to encrypt.");
   const [error, setError] = useState<string>("");
 
@@ -29,11 +25,10 @@ export default function ManualPaymentForm() {
         ts: Date.now(),
       };
 
-      // All the complex logic is now in one function call
-      const encryptedPayload = await encryptClientSide(cardDetails);
+      // All the complex encryption logic is now in one function call
       const encryptedFields = await encryptCardFieldsIndividually(cardDetails);
 
-      console.log({ encryptedFields, encryptedPayload });
+      console.log({ encryptedFields });
 
       setStatus("Data encrypted. Sending to server...");
       const payload = {
@@ -48,13 +43,8 @@ export default function ManualPaymentForm() {
         body: JSON.stringify(payload),
       });
 
-      //   const response = await fetch("/api/process-payment", {
-      //     method: "POST",
-      //     headers: { "Content-Type": "application/json" },
-      //     body: JSON.stringify({ encryptedPayload }),
-      //   });
-
       const result = await response.json();
+
       if (!response.ok) {
         throw new Error(result.error || "An unknown server error occurred.");
       }
@@ -62,6 +52,14 @@ export default function ManualPaymentForm() {
       console.log({ result });
 
       setStatus(`Server response: ${result.message}`);
+
+      // const params = result.response.TxnData.RequestData.paRes;
+      // const sendToOTP = await fetch(
+      //   `${result.response.TxnData.RequestURL}?paRes=${params}`,
+      //   { method: "POST" }
+      // );
+
+      // console.log({ sendToOTP });
     } catch (err: unknown) {
       const message =
         err instanceof Error ? err.message : "An unknown error occurred.";
@@ -73,7 +71,7 @@ export default function ManualPaymentForm() {
 
   return (
     <div style={{ color: "black", maxWidth: "400px" }}>
-      <h3>Educational: Manual Card Encryption (Refactored)</h3>
+      <h3>Manual Card Encryption</h3>
       <form onSubmit={handleSubmit}>
         <input name="pan" placeholder="Card Number" required />
         <input name="cvc" placeholder="CVC" required />
